Moved display to own class
Add two classes and changed code in main.
